:recycle: refactor: 지역이 null 일경우 이름을 뽑지 않도록 수정
