+++ src/main/java/com/sparta/oishitable/domain/common/user/dto/response/UserMyProfileResponse.java
@@ -1,12 +1,11 @@
-package com.sparta.oishitable.domain.common.user.dto.request;
+package com.sparta.oishitable.domain.common.user.dto.response;
 
 import com.sparta.oishitable.domain.common.user.entity.User;
 import lombok.AccessLevel;
 import lombok.Builder;
 
 @Builder(access = AccessLevel.PRIVATE)
-public record UserProfileResponse(
-        String userId,
+public record UserMyProfileResponse(
         String userName,
         String userIntroduce,
         String region,
@@ -14,15 +13,15 @@ public record UserProfileResponse(
         long followingCount
 ) {
 
-    public UserProfileResponse of(User user, long followerCount, long followingCount) {
-        return UserProfileResponse.builder()
-                .userId(user.getId())
+    public static UserMyProfileResponse of(User user, long followerCount, long followingCount) {
+        String region = user.getRegion() != null ? user.getRegion().getName() : null;
+
+        return UserMyProfileResponse.builder()
                 .userName(user.getName())
                 .userIntroduce(user.getIntroduce())
-                .region(user.getRegion().getName())
+                .region(region)
                 .followerCount(followerCount)
                 .followingCount(followingCount)
                 .build();
     }
 }
-
