+++ src/main/java/com/sparta/oishitable/domain/common/user/dto/response/UserProfileResponse.java
@@ -1,11 +1,12 @@
-package com.sparta.oishitable.domain.common.user.dto.request;
+package com.sparta.oishitable.domain.common.user.dto.response;
 
 import com.sparta.oishitable.domain.common.user.entity.User;
 import lombok.AccessLevel;
 import lombok.Builder;
 
 @Builder(access = AccessLevel.PRIVATE)
-public record UserMyProfileResponse(
+public record UserProfileResponse(
+        Long userId,
         String userName,
         String userIntroduce,
         String region,
@@ -13,13 +14,17 @@ public record UserMyProfileResponse(
         long followingCount
 ) {
 
-    public UserMyProfileResponse of(User user, long followerCount, long followingCount) {
-        return UserMyProfileResponse.builder()
+    public static UserProfileResponse of(User user, long followerCount, long followingCount) {
+        String region = user.getRegion() != null ? user.getRegion().getName() : null;
+
+        return UserProfileResponse.builder()
+                .userId(user.getId())
                 .userName(user.getName())
                 .userIntroduce(user.getIntroduce())
-                .region(user.getRegion().getName())
+                .region(region)
                 .followerCount(followerCount)
                 .followingCount(followingCount)
                 .build();
     }
 }
+
